move functions to lib folder
